python-api/src/requirements.txt
psycopg2-binary
python-dotenv
Flask
aws-wsgi